Dropped datatype sizes to save on RAM
Former-commit-id: bb1754c8d447ba4b1c7436a811bc743d07eb76ed

+++ evSeq/util/input_processing.py
@@ -298,7 +298,7 @@ def find_codons_variable_positions(refseq, inframe_bp, plate, well):
             variable_counter += 1
 
     # Convert to an array. Make sure everything is sorted.
-    variable_positions = np.array(variable_positions)
+    variable_positions = np.array(variable_positions, dtype = int)
     variable_positions.sort()
 
     # Make sure that all variable positions are "N" and that we found the appropriate
